+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/protocols/AudioClientProtocol.swift
@@ -10,60 +10,7 @@ import AmazonChimeSDKMedia
 import Foundation
 
 @objc public protocol AudioClientProtocol {
-    // swiftlint:disable function_parameter_count variable_name
-    func startSession(_ host: String!,
-                      basePort port: Int,
-                      callId: String!,
-                      profileId: String!,
-                      microphoneMute mic_mute: Bool,
-                      speakerMute spk_mute: Bool,
-                      isPresenter presenter: Bool,
-                      sessionToken tokenString: String!,
-                      audioWsUrl: String!,
-                      callKitEnabled: Bool) -> audio_client_status_t
-
-    // swiftlint:disable function_parameter_count variable_name
-    func startSession(_ host: String!,
-                      basePort port: Int,
-                      callId: String!,
-                      profileId: String!,
-                      microphoneMute mic_mute: Bool,
-                      speakerMute spk_mute: Bool,
-                      isPresenter presenter: Bool,
-                      sessionToken tokenString: String!,
-                      audioWsUrl: String!,
-                      callKitEnabled: Bool,
-                      appInfo: AppInfo!) -> audio_client_status_t
-
-    // swiftlint:disable function_parameter_count variable_name
-    func startSession(_ host: String!,
-                      basePort port: Int,
-                      callId: String!,
-                      profileId: String!,
-                      microphoneMute mic_mute: Bool,
-                      speakerMute spk_mute: Bool,
-                      isPresenter presenter: Bool,
-                      sessionToken tokenString: String!,
-                      audioWsUrl: String!,
-                      callKitEnabled: Bool,
-                      appInfo: AppInfo!,
-                      audioMode: AudioModeInternal) -> audio_client_status_t
-
-    // swiftlint:disable function_parameter_count variable_name
-    func startSession(_ host: String!,
-                      basePort port: Int,
-                      callId: String!,
-                      profileId: String!,
-                      microphoneMute mic_mute: Bool,
-                      speakerMute spk_mute: Bool,
-                      isPresenter presenter: Bool,
-                      sessionToken tokenString: String!,
-                      audioWsUrl: String!,
-                      callKitEnabled: Bool,
-                      appInfo: AppInfo!,
-                      audioMode: AudioModeInternal,
-                      enableAudioRedundancy: Bool) -> audio_client_status_t
-
+    
     // swiftlint:disable function_parameter_count variable_name
     func startSession(_ host: String!,
                       basePort port: Int,
@@ -78,7 +25,8 @@ import Foundation
                       appInfo: AppInfo!,
                       audioMode: AudioModeInternal,
                       audioDeviceCapabilities: AudioDeviceCapabilitiesInternal,
-                      enableAudioRedundancy: Bool) -> audio_client_status_t
+                      enableAudioRedundancy: Bool,
+                      reconnectTimeoutMs: Int) -> audio_client_status_t
 
     func stopSession() -> Int