+++ AmazonChimeSDKDemo/AmazonChimeSDKDemo/MeetingModel.swift
@@ -34,7 +34,7 @@ class MeetingModel: NSObject {
                                                            logger: logger)
 
     // Utils
-    let logger = ConsoleLogger(name: "MeetingModel")
+    let logger = ConsoleLogger(name: "MeetingModel", level: .INFO)
     let postLogger: PostLogger
     let activeSpeakerObserverId = UUID().uuidString
 
@@ -428,11 +428,13 @@ extension MeetingModel: AudioVideoObserver {
     }
 
     func audioSessionDidDrop() {
+        print("test - audioSessionDidDrop called at \(Date())")
         notifyHandler?("Audio Session Dropped")
         logWithFunctionName()
     }
 
     func audioSessionDidStopWithStatus(sessionStatus: MeetingSessionStatus) {
+        print("test - audioSessionDidStopWithStatus called at \(Date())")
         logWithFunctionName(message: "\(sessionStatus.statusCode)")
 
         removeAudioVideoFacadeObservers()